rm benchmark: fixing signed/unsigned comparison warning.
git-svn-id: http://svn.seqan.de/seqan/trunk@6094 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/benchmarks/read_mappers/apps/compare_sam_wit.h
@@ -317,7 +317,7 @@ compareAlignedReadsToReferenceOnContigForOneRead(Options const & options,
     for (TWitRecordIter it = witRecordsBegin; it != witRecordsEnd; ++it) {
 //         std::cerr << *it << std::endl;
         if (it->isForward != isForward) continue;  // Skip aligned reads on other strand.
-        if (it->distance != options.maxError) continue;  // Skip intervals with wrong distance.
+        if (static_cast<int>(it->distance) != options.maxError) continue;  // Skip intervals with wrong distance.
 
 //         std::cerr << __FILE__ << ":" << __LINE__ << " -- add(" << it->firstPos << ", " << it->lastPos << ", is forward = " << it->isForward << ")" << "  " << *it << std::endl;
         intervalMap[it->lastPos] = FlaggedInterval(it->firstPos, it->lastPos);
